Fix bug
When food was appended to meal, the ID was not being included on the new row. This caused a bug during deletion which caused the totals table to calculate the wrong total, plus the food was not being deleted from the database. Additionally, this fixes a bug where the total goal cal row was not being deleted because it wasn't wrapped in a tr